materialize: minor refactors to lifecycle routines
* Move Transactor to transactor.go (no changes).

* Switch TransactionRequest.Prepared to flow.DriverCheckpoint.

* Drop WriteOpen/WriteOpened, and add WriteAcknowledge.
